Extracted the date range filter type into different repository

+++ app/src/main/java/com/nkuppan/expensemanager/domain/usecase/transaction/DeleteTransactionUseCase.kt
@@ -19,6 +19,10 @@ class DeleteTransactionUseCase @Inject constructor(
             return Resource.Error(Exception("Category type shouldn't be blank"))
         }
 
+        if (transaction.fromAccountId.isBlank()) {
+            return Resource.Error(Exception("From account shouldn't be blank"))
+        }
+
         return repository.deleteTransaction(transaction)
     }
 }
